LazyDistances.PairwiseDistance — Type

PairwiseDistance(f, A, B, [cache_strategy=NullCache()])

Lazily calculate the pairwise result of two arrays A, B with function f.

The output dist::PairwiseDistance is a read-only array type, where dist[i, j] is defined as f(A[i], B[j]).

Each item dist[i, j] is computed lazily during getindex, which could be expensive for non-trivial f. cache_strategy provides a way to automatically store some of the result as cache. The default strategy is NullCache(), which means no cache is enabled.

Warning

How data is cached is internal implementation details and you generally should not talk to cache directly.

Examples

julia> using LazyDistances, Distances

julia> A, B = rand(1:5, 6), rand(1:5, 4);

julia> dist = PairwiseDistance(Euclidean(), A, B)
6×4 PairwiseDistance{Float64, 2}:
 1.0  1.0  2.0  2.0
 4.0  4.0  1.0  1.0
 2.0  2.0  1.0  1.0
 4.0  4.0  1.0  1.0
 0.0  0.0  3.0  3.0
 2.0  2.0  1.0  1.0

LazyDistances.CachedViews.is_cached — Method

is_cached(A::CachedView, i...)::Bool

Check whether the value at index i is cached in A.

This check does not guarantee that cached value is up-to-date.