LazyDistances.PairwiseDistance — Type

PairwiseDistance([index_map], [getindex_op], f, A, B, [cache_strategy])

Lazily calculate the pairwise result of two array-like objects A, B with function f.

The output dist is a read-only array with dist[i, j] defined as f(p, q), where ĩ, j̃ = index_map(i), index_map(j), and p, q = getindex_op(A, ĩ), getindex_op(B, j̃).

Optional Arguments

index_map=identity: index_map is a function-like object that maps given index into a new index or indices.

getindex_op=@view getindex: getindex_op is a funtion-like object that defines how p and q are get. By default it is a viewed version of getindex (without data copy). If getindex_op is provided, then index_map should be provided, too.

cache_strategy=NullCache(): Specify the cache strategy.

Cache Strategies

For non-trivial f, it could be expensive if we calculate it everytime when dist[i, j] is required. cache_strategy provides a way to automatically store some of the result in cache so that f(A[i], B[j]) only gets called once in specific cases.

Current available cache strategies are:

NullCache(): no cache behavior. Every time when you indexing dist[i, j] you're calculating f(A[i], B[j]).

LocalWindowCache(window_size): cache a local window of i so that f(A[i], B[j]) only gets called once if j ∈ i-r:i+r, and gets called multiple times otherwise.

Warning

How data is cached is internal implementation details. Generally, you should not directly talk to cache.

Examples

julia> using LazyDistances, Distances

julia> A, B = 1:6, 1:4

julia> dist = PairwiseDistance(Euclidean(), A, B)
6×4 PairwiseDistance{Float64, 2}:
 0.0  1.0  2.0  3.0
 1.0  0.0  1.0  2.0
 2.0  1.0  0.0  1.0
 3.0  2.0  1.0  0.0
 4.0  3.0  2.0  1.0
 5.0  4.0  3.0  2.0

# For each `i`, `dist[i, i-2:i+2]` will be cached, which means `f(A[i], B[j])` for `j ∈ i-2:i+2`
# will only be calculated once.
julia> dist = PairwiseDistance(Euclidean(), A, B, LocalWindowCache((5, )))
6×4 PairwiseDistance{Float64, 2}:
 0.0  1.0  2.0  3.0
 1.0  0.0  1.0  2.0
 2.0  1.0  0.0  1.0
 3.0  2.0  1.0  0.0
 4.0  3.0  2.0  1.0
 5.0  4.0  3.0  2.0